// src/utilities/types.ts
export interface Config {
  brandName: string;
  assetUrl?: string,
  brandingColorHex?: string;
  backgroundColorHex?: string,
}

export enum ButtonKey {
  Zero = '0',
  One = '1',
  Two = '2',
  Three = '3',
  Four = '4',
  Five = '5',
  Six = '6',
  Seven = '7',
  Eight = '8',
  Nine = '9',
  Clear = 'c',
  Divide = '/',
  Multiply = '*',
  Subtract = '-',
  Add = '+',
  Equal = '=',
}